Address review comments on #266
- Include "Basics.hpp" in headers with no other SKIRT include.
- Convert helper namespaces (NebularLineEmission, NebularContinuumEmission,
  VernerCrossSections, PhotoIonizationRates) to classes with static
  members; move internal helpers and constants into anonymous namespaces
  in the corresponding cpp files.
- Move DIG mix private constant arrays from header to cpp.
- Remove STAB filename user options; hard-code default filenames.
